Fall over safety, sequence verification, success operation vet and test
Installed a health check system so in the even a node goes down the
controller (server) is aware and re-distributes the sequence to another
node. 

Sequences between nodes and between nodes picking up sequences that were
dropped have been verified and tested. 

Success operational parameters have been added, in the event the system
derives the password the information is now relayed to the server and
the issuing client. 

The success information as well as the sequence information should be
stored on a local DB (SQLite) in the case of server failure to prevent
issues. This will need to be implemented.